+++ src/redturtle/volto/interfaces.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
 from plone.restapi.controlpanels.interfaces import IControlpanel
+from plone.base.interfaces.controlpanel import ISiteSchema
 from redturtle.volto import _
 from zope.interface import Interface
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
-from zope.schema import Bool
+from zope.schema import Bool, SourceText, Bytes, Int
+from plone.restapi.controlpanels import IControlpanel
+from plone.autoform import directives as form
 
 
 class IRedturtleVoltoLayer(IDefaultBrowserLayer):
@@ -40,3 +43,64 @@ class IRedTurtleVoltoSettings(Interface):
         default=False,
         required=False,
     )
+
+
+class IRedTurtleVoltoAdditionalSiteSchema(Interface):
+    """
+    Settings interface that add some extra fields to site controlpanel.
+    """
+
+    site_title_translated = SourceText(
+        title=_("site_localized_label", default="Translated site title"),
+        description=_(
+            "site_localized_help",
+            default="If you want to translate site title for different available language, use this field to set translations. If set, this field overrides the default one.",
+        ),
+        required=False,
+        default="",
+    )
+
+    site_subtitle = SourceText(
+        title=_("site_subtitle_label", default="Site subtitle"),
+        description=_(
+            "site_subtitle_help",
+            default="",
+        ),
+        required=False,
+        default="",
+    )
+
+    site_logo_footer = Bytes(
+        title=_("logo_footer_label", default="Footer logo"),
+        description=_(
+            "logo_footer_help",
+            default="Insert a logo that will be used in the site footer.",
+        ),
+        required=False,
+    )
+
+    site_logo_width = Int(required=False)
+    site_logo_height = Int(required=False)
+    site_favicon_width = Int(required=False)
+    site_favicon_height = Int(required=False)
+    site_logo_footer_width = Int(required=False)
+    site_logo_footer_height = Int(required=False)
+
+
+class IRedTurtleVoltoSiteSchema(ISiteSchema, IRedTurtleVoltoAdditionalSiteSchema):
+    """"""
+
+    form.order_after(site_title_translated="site_title")
+    form.order_after(site_subtitle="site_title_translated")
+    form.order_after(site_logo_footer="site_logo")
+
+    form.omitted("site_logo_width")
+    form.omitted("site_logo_height")
+    form.omitted("site_favicon_width")
+    form.omitted("site_favicon_height")
+    form.omitted("site_logo_footer_width")
+    form.omitted("site_logo_footer_height")
+
+
+class IRedTurtleVoltoSiteControlpanel(IControlpanel):
+    """ """